Separate rpc module into client side and server side
